PaimonsNotebook 1.5.2
### 新增

***

- ##### 祈愿记录

  - 祈愿记录设置页新增**导出的祈愿记录**功能
    - 此功能可以对程序导出的祈愿记录进行，删除、修改名称、发送操作

- **成就管理**

  - 成就管理设置页新增**导出的成就记录**功能
    - 此功能可以对程序导出的成就记录进行，删除、修改名称、发送操作

***

### 修改

- ##### 祈愿记录

  - 现在使用祈愿记录导出功能，不再需要外部存储权限
  - 现在导出祈愿记录不再会存储到外部路径，改为存储到程序私有路径
  - **卸载程序后导出的祈愿记录数据会丢失，卸载前确保数据已经转移到其他目录**

- ##### 成就管理

  - 现在使用成就记录导出功能，不再需要外部存储权限
  - 现在导出成就记录不再会存储到外部路径，改为存储到程序私有路径
  - **卸载程序后导出的成就记录数据会丢失，卸载前确保数据已经转移到其他目录**

***

### 修复

- ##### 祈愿记录

  - 修复使用JSON导入祈愿记录时，JSON字段顺序错误导致无法正常导入的问题
  - 修复在安卓12以及上的设备使用祈愿记录导出功能报错的问题
  - 修复导入祈愿记录后，无法立即在切换当前账号中找到导入记录的UID的问题

- ##### 成就管理

  - 修复成就分类详情页锁定交互时提示错误的问题
  - 修复使用JSON导入成就记录时，JSON字段顺序错误导致无法正常导入的问题

+++ app/src/main/java/com/lianyi/paimonsnotebook/common/util/file/FileHelper.kt
@@ -8,10 +8,14 @@ import android.os.Build
 import android.os.Environment
 import android.os.FileUtils
 import android.webkit.MimeTypeMap
+import androidx.core.content.FileProvider
+import com.lianyi.paimonsnotebook.BuildConfig
 import com.lianyi.paimonsnotebook.common.application.PaimonsNotebookApplication
+import com.lianyi.paimonsnotebook.common.extension.number.decimal.format.format
 import com.lianyi.paimonsnotebook.common.util.image.PaimonsNotebookImageLoader
 import java.io.BufferedInputStream
 import java.io.File
+import java.io.FileInputStream
 import java.io.FileOutputStream
 import java.io.InputStream
 import java.net.URI
@@ -21,6 +25,10 @@ object FileHelper {
 
     private val context by lazy {
         PaimonsNotebookApplication.context
+    }
+
+    val contentResolver by lazy {
+        context.contentResolver
     }
 
     private val privateStoragePath by lazy {
@@ -55,14 +63,14 @@ object FileHelper {
             privateStoragePath?.resolve("package")!!
 
     //存储祈愿记录的路径
-    private val saveFileGachaItemsPath
+    val saveFileGachaItemsPath
         get() =
-            rootPath?.resolve("gacha")!!
+            privateStoragePath?.resolve("gacha")!!
 
     //存储成就记录的路径
-    private val saveFileAchievementsPath
+    val saveFileAchievementsPath
         get() =
-            rootPath?.resolve("achievements")!!
+            privateStoragePath?.resolve("achievements")!!
 
 
     //扫描图片,使其出现在相册中
@@ -123,7 +131,6 @@ object FileHelper {
 
         if (file.exists()) {
             file.delete()
-            file.mkdirs()
         }
 
         FileOutputStream(file).use {
@@ -164,10 +171,14 @@ object FileHelper {
         } else null
 
     fun uriToFile(uri: Uri): File? {
-        return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
-            uriToFileInAndroidQ(uri)
-        } else {
-            File(URI(uri.toString()))
+        return try {
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
+                uriToFileInAndroidQ(uri)
+            } else {
+                File(URI(uri.toString()))
+            }
+        } catch (_: Exception) {
+            null
         }
     }
 
@@ -189,6 +200,25 @@ object FileHelper {
         return File(rootPath, fileName)
     }
 
+    /*
+    * 获取文件uri
+    * scheme = content
+    * */
+    fun getContentUriForFile(file: File) =
+        FileProvider.getUriForFile(context, "${BuildConfig.APPLICATION_ID}.provider", file)
+
+    /*
+    * 从content uri获取file out put stream
+    * */
+    fun getFileOutputStreamByContentUri(uri: Uri, mode: String = "rw") =
+        contentResolver.openFileDescriptor(uri, mode)?.use { fileDescriptor ->
+            FileOutputStream(fileDescriptor.fileDescriptor)
+        }
+
+    fun getFileInputStreamByContentUri(uri: Uri, mode: String = "rw") =
+        contentResolver.openFileDescriptor(uri, mode)?.use { fileDescriptor ->
+            FileInputStream(fileDescriptor.fileDescriptor)
+        }
 
     /*
     * 以流的方式保存文件
@@ -244,7 +274,12 @@ object FileHelper {
         val file = File(rootPath, fileName)
 
         if (file.exists() && onExistsDelete) {
-            file.delete()
+            val uri = getContentUriForFile(file)
+
+            val deleteRow = contentResolver.delete(uri, null, null)
+            if (deleteRow == 0) {
+                file.delete()
+            }
         }
 
         return file
@@ -280,6 +315,31 @@ object FileHelper {
             }
         }
     }
+
+
+    /*
+    * 获得文件尺寸描述文本
+    * */
+    fun getFileSizeDescribeString(length: Long): String {
+        val kb = byteToKB(length)
+
+        if (kb < 1024) {
+            return "${kb.format()}KB"
+        }
+
+        val mb = byteToMB(length)
+
+        if (mb < 1024) {
+            return "${mb.format()}MB"
+        }
+
+        return "${byteToGB(length).format()}GB"
+    }
+
+    private fun byteToKB(length: Long) = length / 1024f
+    private fun byteToMB(length: Long) = length / (1024f * 1024f)
+    private fun byteToGB(length: Long) = length / (1024f * 1024f * 1024)
+
 
     private val df
         get() =
